requirements.txt
boto3==1.13.10
botocore==1.16.10
docutils==0.15.2
jmespath==0.10.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
termcolor==1.1.0
urllib3==1.25.9
python-dotenv==0.14.0
